backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-doc=0.0.4=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.12.1=pypi_0
bcrypt=4.0.1=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.12.2=haa95532_0
cachetools=6.2.6=pypi_0
certifi=2026.2.25=pypi_0
cffi=2.0.0=pypi_0
charset-normalizer=3.4.4=pypi_0
click=8.3.1=pypi_0
colorama=0.4.6=pypi_0
cryptography=46.0.5=pypi_0
deprecation=2.1.0=pypi_0
distro=1.9.0=pypi_0
dnspython=2.8.0=pypi_0
ecdsa=0.19.1=pypi_0
email-validator=2.3.0=pypi_0
faiss-cpu=1.13.2=pypi_0
fastapi=0.131.0=pypi_0
filelock=3.24.3=pypi_0
fsspec=2026.2.0=pypi_0
groq=1.0.0=pypi_0
h11=0.16.0=pypi_0
h2=4.3.0=pypi_0
hf-xet=1.3.2=pypi_0
hpack=4.1.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=1.5.0=pypi_0
hyperframe=6.1.0=pypi_0
idna=3.11=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.3=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
langchain-core=1.2.16=pypi_0
langchain-text-splitters=1.1.1=pypi_0
langsmith=0.7.9=pypi_0
libexpat=2.7.4=hd7fb8db_0
libffi=3.4.4=hd77b12b_1
libmpdec=4.0.0=h827c3e9_0
libzlib=1.3.1=h02ab6af_0
lz4-c=1.9.4=h2bbff1b_1
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.3=pypi_0
mdurl=0.1.2=pypi_0
mmh3=5.2.0=pypi_0
motor=3.7.1=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.7.1=pypi_0
networkx=3.6.1=pypi_0
numpy=2.4.2=pypi_0
openssl=3.0.19=hbb43b14_0
orjson=3.11.7=pypi_0
packaging=25.0=py314haa95532_1
passlib=1.7.4=pypi_0
pip=26.0.1=pyhc872135_0
postgrest=2.28.0=pypi_0
propcache=0.4.1=pypi_0
pyasn1=0.6.2=pypi_0
pycparser=3.0=pypi_0
pydantic=2.12.5=pypi_0
pydantic-core=2.41.5=pypi_0
pydantic-settings=2.13.1=pypi_0
pygments=2.19.2=pypi_0
pyiceberg=0.11.0=pypi_0
pyjwt=2.11.0=pypi_0
pymongo=4.16.0=pypi_0
pyparsing=3.3.2=pypi_0
pypdf=6.7.4=pypi_0
pyroaring=1.0.3=pypi_0
python=3.14.2=h3bc0a5a_101_cp314
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.2.1=pypi_0
python-jose=3.5.0=pypi_0
python-multipart=0.0.22=pypi_0
python_abi=3.14=2_cp314
pyyaml=6.0.3=pypi_0
realtime=2.28.0=pypi_0
regex=2026.2.28=pypi_0
requests=2.32.5=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=14.3.3=pypi_0
rsa=4.9.1=pypi_0
safetensors=0.7.0=pypi_0
scikit-learn=1.8.0=pypi_0
scipy=1.17.1=pypi_0
sentence-transformers=5.2.3=pypi_0
setuptools=80.10.2=py314haa95532_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.51.1=hda9a48d_0
starlette=0.52.1=pypi_0
storage3=2.28.0=pypi_0
strenum=0.4.15=pypi_0
strictyaml=1.7.3=pypi_0
supabase=2.28.0=pypi_0
supabase-auth=2.28.0=pypi_0
supabase-functions=2.28.0=pypi_0
sympy=1.14.0=pypi_0
tenacity=9.1.4=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.15=hf199647_0
tokenizers=0.22.2=pypi_0
torch=2.10.0=pypi_0
tqdm=4.67.3=pypi_0
transformers=5.2.0=pypi_0
typer=0.24.1=pypi_0
typer-slim=0.24.0=pypi_0
typing-extensions=4.15.0=pypi_0
typing-inspection=0.4.2=pypi_0
tzdata=2025c=he532380_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.6.3=pypi_0
uuid-utils=0.14.1=pypi_0
uvicorn=0.41.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
websockets=15.0.1=pypi_0
wheel=0.46.3=py314haa95532_0
xxhash=3.6.0=pypi_0
xz=5.6.4=h4754444_1
yarl=1.22.0=pypi_0
zlib=1.3.1=h02ab6af_0
zstandard=0.25.0=pypi_0
zstd=1.5.7=h56299aa_0
